+++ src/main/java/com/cisowski/schoolmanagement/users/teacher/repository/TeacherRepository.java
@@ -8,12 +8,18 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
+import java.util.Optional;
 
 @Repository
 public interface TeacherRepository extends BaseUserRepository<TeacherEntity, Integer> {
-    boolean existsByTeachingSubjects(SubjectEntity subject);
-    List<TeacherEntity> findByTeachingSubjects(SubjectEntity subject);
 
-    @Query("SELECT s FROM SubjectEntity s JOIN s.teachers t WHERE t.id = :teacherId")
-    List<SubjectEntity> findByTeacherId(@Param("teacherId") Integer teacherId);
+    boolean existsByTeachingSubjectsAndIsHideFalse(SubjectEntity subject);
+    List<TeacherEntity> findByTeachingSubjectsAndIsHideFalse(SubjectEntity subject);
+
+    @Query("SELECT s FROM SubjectEntity s JOIN s.teachers t WHERE t.id = :teacherId AND t.isHide = false")
+    List<SubjectEntity> findByTeacherIdAndIsHideFalse(@Param("teacherId") Integer teacherId);
+
+    Optional<TeacherEntity> findByIdAndIsHideFalse(Integer teacherId);
+
+    List<TeacherEntity> findAllByIsHideFalse();
 }