+++ src/main/resources/db/migration/V24__add_user_hide_flag.sql
@@ -1,0 +1,3 @@
+ALTER TABLE users
+ADD is_hide bit NOT NULL
+DEFAULT 0;