+++ src/test/java/com/cisowski/schoolmanagement/integration/BaseIntegrationTest.java
@@ -32,7 +32,6 @@ import java.util.Random;
 import static io.restassured.RestAssured.given;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@Testcontainers
 @ActiveProfiles("test")
 public abstract class BaseIntegrationTest {